streaming is working - time to merge in
